requirements.txt
# Core Framework
fastapi>=0.103.0,<0.104.0
uvicorn>=0.23.0,<0.24.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
asyncpg>=0.28.0,<0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9,<3.0.0  # PostgreSQL sync driver

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
firebase-admin>=6.2.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0

# Data Processing & Analysis
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# HTTP & Networking
requests>=2.31.0,<3.0.0
aiohttp>=3.8.4,<4.0.0
httpx>=0.25.0,<0.26.0  # Modern HTTP client with async support
googletrans>=3.1.0a0  # Updated to newer version that's compatible with modern httpx

# Caching
redis>=5.0.0,<6.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-cov>=4.1.0,<5.0.0
black>=23.9.0,<24.0.0
flake8>=6.1.0,<7.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.0,<2.0.0

# Monitoring & Logging
prometheus-client>=0.17.0,<0.18.0
python-json-logger>=2.0.7,<3.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.4.0,<10.0.0

# Web Scraping
beautifulsoup4>=4.12.0,<5.0.0
selenium>=4.12.0,<5.0.0

# Date/Time handling
pytz>=2023.3
python-dateutil>=2.8.2,<3.0.0

# Performance & Optimization
ujson>=5.8.0,<6.0.0  # Faster JSON processing
orjson>=3.9.0,<4.0.0  # Even faster JSON processing
